kernel: owl-loader: load the EEPROM firmware synchronously
This is needed because otherwise there may be a race-condition with
userspace, because "wifi detect" is called before the OWL-fixup has
completed. "wifi detect" relies on all wireless interfaces being
available once it is called.
However, the ath9k wireless interface is only created after the
OWL-fixup has completed, because before the PCI ID does not match
ath9k.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>